refactor: clean up copy.ts TDE-1551
